Fix: 에러페이지 빌드 실패 로직 수정
